+++ models/Mrecipe.js
@@ -50,7 +50,7 @@ const RecipesModel = (sequelize, DataTypes) =>{
     },
     { // param3 : 모델 옵션 정의
         freezeTableName : true , // 테이블 명을 그대로 사용한다. (복수형으로 바까주지 X)
-        timestamps : false, // 데이터의 추가/수정 시간을 자동으로 기록(컬럼)
+        timestamps : true, // 데이터의 추가/수정 시간을 자동으로 기록(컬럼)
     }
 );
     return Recipes;
